add formBlog
