+++ src/app/globals.css
@@ -1,5 +1,5 @@
-@import "tailwindcss";
 @import url("https://fonts.googleapis.com/css2?family=Cormorant:ital,wght@0,300..700;1,300..700&display=swap");
+@import "tailwindcss";
 
 @theme inline {
   --font-sans:
